src/requirements.txt
Jinja2==3.1.2
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
